Lower start timeout; raise finish timeout
I don't think there's much that affects the time it takes for a Pages
build to start, but the Pages build itself is quite variable, it turns,
especially with large amounts of data being copied across and then
deployed. Extrapolating from a pipeline I saw I think 1GB of data would
take about 1.5min. I figure that so long as the build starts and is
acquired, it doesn't really matter how long it takes.